requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
Fabric==1.8.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==18.0
Jinja2==2.10.1
SQLAlchemy==1.3
coverage==3.7.1
paramiko==1.12.2
pycrypto==2.6.1
Flask-WTF==0.9.4
pytest-xdist==1.10
python-dotenv==0.16
dataclasses==0.8
uwsgi==2.0.19.1
newrelic==6.2.0.156